// src/generated/ts-rs/ServersEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Company } from "./Company";
import type { Credential } from "./Credential";
import type { ErasedService } from "./ErasedService";

export type ServersEvent = { "Updated": Array<Company<ErasedService>> } | { "Expanded": Array<string> } | { "ServiceCredential": { service: string, remember: boolean, credential: Credential | null, } } | "Empty";